+++ totp/cli/cli_helpers.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <cli/cli.h>
-#include "../../types/plugin_state.h"
+#include "../types/plugin_state.h"
 
 #define TOTP_CLI_COMMAND_NAME "totp"
