+++ totp/types/plugin_state.h
@@ -4,7 +4,7 @@
 #include <gui/gui.h>
 #include <dialogs/dialogs.h>
 #include "../services/list/list.h"
-#include "../scenes/totp_scenes_enum.h"
+#include "../ui/totp_scenes_enum.h"
 
 #define TOTP_IV_SIZE 16
